refactor(file-structure): move grid to different file
