+++ src/terminal_pane/mod.rs
@@ -1,7 +1,7 @@
-mod scroll;
+mod grid;
 mod terminal_character;
 mod terminal_pane;
 
-pub use scroll::*;
+pub use grid::*;
 pub use terminal_character::*;
 pub use terminal_pane::*;
